ValueObservation combination
